requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
google-generativeai==0.8.6
python-dotenv==1.2.1
Pillow==11.3.0
requests==2.32.4
